backend/LLM_Score/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.26.0
python-dotenv>=1.0.0
firebase-admin>=6.2.0
openai>=1.40.0
